feat [DEV-27]: Add storybook and refactor linters structure
