+++ apps/frontend/angular/platform/eslint.config.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @nx/enforce-module-boundaries
-import angularConfig from '../../../../eslint.config.angular.js';
+import angularConfig from '../../../../linters/eslint/configs/angular.config.js';
 
 /** @type { import("eslint").Linter.Config[] } */
 export default angularConfig;
