+++ libs/angular/ui/eslint.config.js
@@ -1,5 +1,7 @@
 // eslint-disable-next-line @nx/enforce-module-boundaries
-import angularConfig from '../../../eslint.config.angular.js';
+import angularConfig from '../../../linters/eslint/configs/angular.config.js';
+// eslint-disable-next-line @nx/enforce-module-boundaries
+import storybookConfig from '../../../linters/eslint/configs/storybook.config.js';
 
 /** @type { import("eslint").Linter.Config[] } */
-export default angularConfig;
+export default [...angularConfig, ...storybookConfig];
