+++ libs/angular/ui/project.json
@@ -37,6 +37,55 @@
 			"options": {
 				"lintFilePatterns": ["./libs/angular/ui/src/**/*.scss"]
 			}
+		},
+		"storybook": {
+			"executor": "@storybook/angular:start-storybook",
+			"options": {
+				"port": 4400,
+				"configDir": "libs/angular/ui/.storybook",
+				"browserTarget": "ng-ui:build-storybook",
+				"compodoc": false
+			},
+			"configurations": {
+				"ci": {
+					"quiet": true
+				}
+			}
+		},
+		"build-storybook": {
+			"executor": "@storybook/angular:build-storybook",
+			"outputs": ["{options.outputDir}"],
+			"options": {
+				"outputDir": "dist/storybook/ng-ui",
+				"configDir": "libs/angular/ui/.storybook",
+				"browserTarget": "ng-ui:build-storybook",
+				"compodoc": false
+			},
+			"configurations": {
+				"ci": {
+					"quiet": true
+				}
+			}
+		},
+		"test-storybook": {
+			"executor": "nx:run-commands",
+			"options": {
+				"command": "test-storybook -c libs/angular/ui/.storybook --url=http://localhost:4400"
+			}
+		},
+		"static-storybook": {
+			"executor": "@nx/web:file-server",
+			"dependsOn": ["build-storybook"],
+			"options": {
+				"buildTarget": "ng-ui:build-storybook",
+				"staticFilePath": "dist/storybook/ng-ui",
+				"spa": true
+			},
+			"configurations": {
+				"ci": {
+					"buildTarget": "ng-ui:build-storybook:ci"
+				}
+			}
 		}
 	}
 }
