+++ libs/angular/ui/tsconfig.json
@@ -8,6 +8,9 @@
 		},
 		{
 			"path": "./tsconfig.spec.json"
+		},
+		{
+			"path": "./.storybook/tsconfig.json"
 		}
 	]
 }
