+++ libs/angular/ui/tsconfig.lib.json
@@ -1,5 +1,12 @@
 {
 	"extends": "./tsconfig.json",
-	"exclude": ["src/**/*.spec.ts", "src/test-setup.ts", "jest.config.ts", "src/**/*.test.ts"],
+	"exclude": [
+		"src/**/*.spec.ts",
+		"src/test-setup.ts",
+		"jest.config.ts",
+		"src/**/*.test.ts",
+		"**/*.stories.ts",
+		"**/*.stories.js"
+	],
 	"include": ["src/**/*.ts"]
 }
