+++ package.json
@@ -65,10 +65,18 @@
 		"@nx/jest": "20.3.2",
 		"@nx/js": "20.3.2",
 		"@nx/playwright": "20.3.2",
+		"@nx/storybook": "20.3.2",
 		"@nx/web": "20.3.2",
 		"@nx/workspace": "20.3.2",
 		"@playwright/test": "^1.49.1",
 		"@schematics/angular": "19.0.6",
+		"@storybook/addon-essentials": "^8.4.6",
+		"@storybook/addon-interactions": "^8.4.6",
+		"@storybook/angular": "^8.4.6",
+		"@storybook/core-server": "^8.4.6",
+		"@storybook/jest": "^0.2.3",
+		"@storybook/test-runner": "^0.19.0",
+		"@storybook/testing-library": "^0.2.2",
 		"@swc-node/register": "~1.10.9",
 		"@swc/core": "~1.10.1",
 		"@swc/helpers": "~0.5.15",
@@ -77,6 +85,7 @@
 		"autoprefixer": "^10.4.20",
 		"esbuild": "^0.19.2",
 		"eslint": "~9.17.0",
+		"eslint-plugin-storybook": "^0.11.2",
 		"husky": "^9.1.7",
 		"jest": "^29.7.0",
 		"jest-environment-jsdom": "^29.7.0",
@@ -88,6 +97,7 @@
 		"nx": "20.3.2",
 		"nx-stylelint": "^18.0.0",
 		"prettier": "^3.4.2",
+		"storybook": "^8.4.6",
 		"stylelint": "^16.12.0",
 		"ts-jest": "^29.2.5",
 		"ts-node": "10.9.2",
